Add files via upload
With car prices rising faster than Gen-Z’s mental health decline factor,
This is a machine learning model built to predict the prices of used cars. Data on the prices of used cars over the years from 2003 to 2018. Taking into account all the factors that the price of cars depends upon, like the brand of the car and features such as horsepower and mileage, I have learned how to train a car price prediction model.

Plotting various graphs using these features, we learned that the selling price of the cars was the highest in 2017 and the second highest in 2010.  After scaling transformations are performed, ensuring all the features contribute equally to the model, a histogram of scaled prices is plotted.